faster-whisper/requirements.txt
# Core ML Libraries
torch>=2.1.0,<2.3.0
torchaudio>=2.1.0,<2.3.0
torchvision>=0.16.0,<0.18.0
transformers>=4.35.0
datasets>=2.14.0
evaluate>=0.4.0
jiwer>=3.0.0

# Lightning AI
lightning>=2.1.0
torchmetrics>=1.2.0,<1.3.0

# Efficient Training
deepspeed>=0.11.0
peft>=0.6.0
bitsandbytes>=0.41.0

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0
ctranslate2>=3.20.0
faster-whisper>=0.9.0

# Gradio UI
gradio[oauth]>=4.14.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# Logging and Monitoring
wandb>=0.15.0
tensorboard>=2.14.0

# Utilities
tqdm>=4.66.0
python-dotenv>=1.0.0
pyyaml>=6.0

# Audio Augmentation
audiomentations>=1.4.0
torch-audiomentations>=0.11.0

# Voice Activity Detection
silero-vad>=4.0.0

# Model Optimization
onnx>=1.15.0
onnxruntime-gpu>=1.16.0

# UI Enhancements
plotly>=5.18.0

# Hyperparameter Optimization
optuna>=3.5.0

# API Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Monitoring
prometheus-client>=0.19.0
sentry-sdk>=1.40.0

# WebSocket Support
websockets>=12.0

# Amharic Text Processing
unicodedataplus>=0.0.1
ethiopic>=0.8.0